refactor: Refatora o controller para incluir o responseEntity nas minhas rotas e customizar a resposta das requisições do cliente

+++ src/main/java/dev/Dam/Cadastro/Pessoas/Controller/Service/UserController.java
@@ -1,5 +1,7 @@
 package dev.Dam.Cadastro.Pessoas.Controller.Service;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,36 +23,60 @@ public class UserController {
 
     // Adicionar usuário (CREATE)
     @PostMapping("/criar")
-    public UserDTO criateUser(@RequestBody UserDTO user) {
-        return userService.createUser(user);
+    public ResponseEntity<String> createUser(@RequestBody UserDTO user) {
+        UserDTO newUser = userService.createUser(user);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Usuário criado com sucesso: " + newUser.getName() + " (ID): " + newUser.getId());
+
     }
 
     // Mostrar todos os usuários (READ)
     @GetMapping("/listar")
-    public List<UserDTO> listUsers() {
-        return userService.listUsers();
+    public ResponseEntity<List<UserDTO>> listUsers() {
+        List<UserDTO> users = userService.listUsers();
+        return ResponseEntity.ok(users);
     }
 
     //localhost
     // Procurar usuário por ID (READ)
     @GetMapping("/listar/{id}")
-    public UserDTO showAllUsersByID(@PathVariable Long id) {
-        return userService.listUserByID(id);
+    public ResponseEntity<?> showUsersByID(@PathVariable Long id) {
+        UserDTO user = userService.listUserByID(id);
+        if (user != null) {
+            return ResponseEntity.ok(user);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Usuário de id: " + id + ", não existe.");
+        }
     }
+
 
     // Alterar dados dos usuários (UPDATE)
     // Sempre que o usuario for passar uma variável, tem que isar chaves
     //E também o @PathVariable
     @PutMapping("/alterar/{id}")
-    public UserDTO changeUsersByID(@PathVariable Long id, @RequestBody UserDTO updatedUser) {
-        return userService.updateUser(id, updatedUser);
+    public ResponseEntity<?> changeUsersByID(@PathVariable Long id, @RequestBody UserDTO updatedUser) {
+        UserDTO user = userService.updateUser(id, updatedUser);
+        if(user != null) {
+            return ResponseEntity.ok(user);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Usuário de id: " + id + ", não existe.");
+        }
     }
 
     // Deletar usuários (DELETE)
     //@PathVariable - Serve para colocar uma variávio que o usuário deseja inserir
     @DeleteMapping("/deletar/{id}")
-    public void deleteUsersByID(@PathVariable Long id) {
-        userService.deleteUsersByID(id);
+    public ResponseEntity<String> deleteUsersByID(@PathVariable Long id) {
+        if(userService.listUserByID(id) != null) {
+            userService.deleteUsersByID(id);
+            return ResponseEntity.ok("Usuário do ID " + id + ", foi deletado.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("O usuário com o ID " + id + " não existe.");
+        }
+
     }
 
 }
